+++ coders/map.c
@@ -452,7 +452,7 @@ static MagickBooleanType WriteMAPImage(const ImageInfo *image_info,Image *image)
     {
       if (image->colors > 256)
         *q++=(unsigned char) ((size_t) GetPixelIndex(indexes+x) >> 8);
-      *q++=(unsigned char) GetPixelIndex(indexes+x);
+      *q++=(unsigned char) ((size_t) GetPixelIndex(indexes+x));
     }
     (void) WriteBlob(image,(size_t) (q-pixels),pixels);
   }
