+++ coders/viff.c
@@ -1182,7 +1182,7 @@ static MagickBooleanType WriteVIFFImage(const ImageInfo *image_info,
               break;
             indexes=GetVirtualIndexQueue(image);
             for (x=0; x < (ssize_t) image->columns; x++)
-              *q++=(unsigned char) GetPixelIndex(indexes+x);
+              *q++=(unsigned char) ((size_t) GetPixelIndex(indexes+x));
             if (image->previous == (Image *) NULL)
               {
                 status=SetImageProgress(image,SaveImageTag,(MagickOffsetType) y,
